Moved GetInTouch files into Home folder
